fix(app): widen _battery_score wired/wireless params to float | None
Follow-up to 66c8d34: scoring._battery_score still declared the charging
wattage parameters as `int | None`, so mypy fails on the call site that
passes `smartphone.charging_wired_w / charging_wireless_w` (now float).

Refs #1

+++ app/services/scoring.py
@@ -101,8 +101,8 @@ def _camera_score(cameras: list[dict[str, Any]]) -> float | None:
 
 def _battery_score(
     battery_mah: int,
-    wired_w: int | None,
-    wireless_w: int | None,
+    wired_w: float | None,
+    wireless_w: float | None,
     process_nm: float | None,
 ) -> float | None:
     if battery_mah <= 0:
